src/handlers/DeleteThingItem/requirements.txt
-e src/common/
aws_lambda_powertools
boto3-stubs[dynamodb]
